requirements.txt
asgiref==3.5.0
autopep8==1.6.0
biopython==1.79
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.2
django-heroku==0.3.1
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
filelock==3.4.2
gunicorn==20.1.0
numpy==1.22.4
Pillow==9.2.0
platformdirs==2.4.1
psycopg2==2.9.3
pycodestyle==2.8.0
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
tzdata==2021.5
virtualenv==20.13.0
whitenoise==6.1.0
